Add dhfind logic to dhstore
DHFind handling can be enabled by specifying a providers URL (--providersURL flag). When enabled, this causes dhstore to handle a multihash/ or a cid/ GET requests as the dhfind service handles it, when a CID or non-dh-multihash is received. This includes support for streaming nd-json. When dhstore receives a multihash/ GET with a dh-multihash it operates as dhstore did previously, even if dhfind is enabled.

This eliminates the need for a separate dhfind service, and provides dhfind functionality without additional network hops.

Fixes #52